Back to pyarrow
Fastparquet doesn't encode categorical columns in a way that the the arrow R
package can understand, and the file size is twice as large. I believe the memory
issue that precipitated the change was due to an edge case in cruise filtering of
GP15_2 that shouldn't come up again with correct filtering parameters.