Add a --delta flag to filtering to only filter new files
As the code existed before this any time an hourly OPP parquet file was written it erased
any file at the same path, which is fine for batch processing but isn't ideal for realtime
filtering where new data could be filtered every 3 minutes.

This required a new feature that merges a new OPP parquet with any existing OPP file at
the same path. Old matching entries are overwritten by new entries. Old entries that don't
exist in the new data are kept unchanged. And new data is saved as normal.