Create separate serial and parallel filtering paths, fix bugs
There are two bugs this commit fixes. The case where an EVT file an
EVT file could not be read was not handled properly when checking
for maximum particle count during filter. And there was shared state
between hourly filtering groups for errors and results lists that
only exists in the serial code path. For parallel filtering these
lists are pickled first and aren't the same objects for all jobs,
but in the serial case they are the same object which leads to
incorrect results and runtime errors. This bug would not have
affected previous versions of seaflowpy since the serial path did
not exist.