Fix `evt parquet` bug and improve `evt validate` hash
`evt parquet` got files mixed up between LabView binary inputs
and parquet outputs if the SFL and EVT file sets weren't exactly
matched. This has been fixed in this commit.

The hash used in `evt validate` and `evt compare` depended on the
specific serialization of a pandas dataframe, which changed in
pandas 2.2.0. This commit switches to using the MD5 checksum of
the bytes representation of the underlying numpy matrix.