Add code to "curate" VCT data
The first step in using VCT data is to select a single quantile,
select Qc and diam with the appropriate index of refraction for
each population, and remove outlier time points. The new vct.curate
function and `vct curate` CLI captures this process and should make
it easier to produce a ready to use version of VCT files.